Refactor: removing a crash from `sku`
